Added screen change anims, bag sounds, tweaked.
Added bag sounds;
Added BasicScreen: implements event-based screen change scheme, screen
fade-in and fade-out animations;
Refactored screen classes to utilize the new 'onEnter' and 'onReady'
overrides. The onEnter() is called when the screen is about to be added
to stage, before the fade-in animation, and after that animation the
onReady() is called. The onExit() is called right after the fade-out
animation, before removing from stage.